+++ CHANGELOG.md
@@ -1,8 +1,12 @@
+## [0.2.1] - 2022-11-05
+
+- Add `String.ANSI` module
+
 ## [0.2.0] - 2022-10-28
 
 - Drop the `Data` prefix from `Data.Text.ANSI` and `Data.Text.Builder.ANSI` modules
 - Rename `Text.Builder.ANSI` to `Text.Lazy.Builder.ANSI`
-- Add `Text.Builder.ANSI` back, but based on `text-builder`'s builder instead of `text`'s builder.
+- Add `Text.Builder.ANSI` back, but based on `text-builder`'s builder instead of `text`'s builder
 
 ## [0.1.1] - 2021-01-07